refactor: Move database tests to database.rs and add from_env test
- Move test_database_migrations from lib.rs to database.rs for better organization
- Add test_database_from_env to verify Database::from_env() functionality
- Clean up lib.rs to contain only module declarations and exports
- Both tests use testcontainers for isolated testing environment
